+++ src/CredoProvider.ts
@@ -107,7 +107,8 @@ export default class CredoProvider {
   private parse(output: string): CredoOutput | null {
     if (output.length < 1) {
       vscode.window.showWarningMessage(
-        `command \`${this.config.command} credo\` returns empty output! please check configuration.`,
+        `command \`${this.config.command} credo\` returns empty output! please check configuration.
+        Did you add or modify your dependencies? You might need to run \`mix deps.get\` or recompile.`,
       );
 
       return null;